Переменные переведены в double, для точного определения. Обновлен счетчик времени

+++ src/com/company/Bicycle.java
@@ -2,7 +2,7 @@ package com.company;
 
 public class Bicycle extends Vehicle {
 
-    public Bicycle(String name,int speed, int rts, int priceperkm) {
+    public Bicycle(String name,double speed, double rts, double priceperkm) {
         super(name,speed, rts, priceperkm);
         rest = 0.3;
         restday = 8;
