+++ src/com/company/Car.java
@@ -2,7 +2,7 @@ package com.company;
 
 public class Car extends Vehicle {
 
-    public Car(String name,int speed, int rts, int priceperkm) {
+    public Car(String name,double speed, double rts, double priceperkm) {
         super(name,speed, rts, priceperkm);
         rest = 0.5;
         restday = 7;
