+++ src/com/company/Wagon.java
@@ -2,7 +2,7 @@ package com.company;
 
 public class Wagon extends Vehicle {
 
-    public Wagon(String name,int speed, int rts, int priceperkm) {
+    public Wagon(String name,double speed, double rts, double priceperkm) {
         super(name,speed, rts, priceperkm);
         rest = 1;
         restday = 10;
